Split classes up into files
