+++ lib/sidedock/base.rb
@@ -1,0 +1,11 @@
+module Sidedock
+  class Base
+    def machine
+      self.class.machine
+    end
+
+    def self.machine
+      @@machine ||= Machine.new 'sidedock'
+    end
+  end
+end
