+++ lib/sidedock/cli/docker_cli.rb
@@ -1,0 +1,3 @@
+class DockerCLI < CLI
+  self.program = 'docker'
+end
