+++ lib/sidedock/cli/machine_cli.rb
@@ -1,0 +1,3 @@
+class MachineCLI < CLI
+  self.program = 'docker-machine'
+end
